feat: add InventoryStartedDetailsViewAllContainer
